move components to summary/
