+++ src/components/summary/Featured.svelte
